Separate out jetty websocket api dependency into a separate maven module.
Tweak the packaging scripts to package the websocket dependencies correctly.
