+++ agent/test/unit/com/thoughtworks/go/agent/AgentControllerTest.java
@@ -33,12 +33,12 @@ import com.thoughtworks.go.remote.BuildRepositoryRemote;
 import com.thoughtworks.go.remote.work.Work;
 import com.thoughtworks.go.server.service.AgentBuildingInfo;
 import com.thoughtworks.go.server.service.AgentRuntimeInfo;
-import com.thoughtworks.go.server.websocket.Action;
-import com.thoughtworks.go.server.websocket.Message;
 import com.thoughtworks.go.util.SubprocessLogger;
 import com.thoughtworks.go.util.SystemEnvironment;
 import com.thoughtworks.go.util.SystemUtil;
 import com.thoughtworks.go.util.command.EnvironmentVariableContext;
+import com.thoughtworks.go.websocket.Action;
+import com.thoughtworks.go.websocket.Message;
 import com.thoughtworks.go.work.SleepWork;
 import org.junit.After;
 import org.junit.Before;
