+++ common/pom.xml
@@ -155,11 +155,6 @@
                 </exclusion>
             </exclusions>
         </dependency>
-        <dependency>
-            <groupId>org.eclipse.jetty.websocket</groupId>
-            <artifactId>websocket-api</artifactId>
-            <version>9.2.11.v20150529</version>
-        </dependency>
     </dependencies>
 
     <build>
