+++ jetty9-websocket/src/com/thoughtworks/go/websocket/JsonMessage.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 ThoughtWorks, Inc.
+ * Copyright 2016 ThoughtWorks, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.thoughtworks.go.server.websocket;
+package com.thoughtworks.go.websocket;
 
 import com.google.gson.*;
 import com.thoughtworks.go.domain.FetchHandler;
