+++ server/src/com/thoughtworks/go/server/service/BuildAssignmentService.java
@@ -25,11 +25,11 @@ import com.thoughtworks.go.remote.work.*;
 import com.thoughtworks.go.server.materials.StaleMaterialsOnBuildCause;
 import com.thoughtworks.go.server.service.builders.BuilderFactory;
 import com.thoughtworks.go.server.transaction.TransactionTemplate;
-import com.thoughtworks.go.server.websocket.Action;
 import com.thoughtworks.go.server.websocket.Agent;
 import com.thoughtworks.go.server.websocket.AgentRemoteHandler;
-import com.thoughtworks.go.server.websocket.Message;
 import com.thoughtworks.go.util.TimeProvider;
+import com.thoughtworks.go.websocket.Action;
+import com.thoughtworks.go.websocket.Message;
 import org.apache.commons.collections.Closure;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
