+++ server/src/com/thoughtworks/go/server/websocket/Agent.java
@@ -16,6 +16,8 @@
 
 package com.thoughtworks.go.server.websocket;
 
+import com.thoughtworks.go.websocket.Message;
+
 import java.util.concurrent.Future;
 
 public interface Agent {
