+++ tasks/jars.rake
@@ -36,8 +36,13 @@ def server_launcher_dependencies
   jars = Dir.glob($PROJECT_BASE + '/server-launcher/target/libs/*.jar')
   jars
 end
+
 def jetty_jars
   Dir.glob($PROJECT_BASE + '/jetty9/target/lib/*.jar')
+end
+
+def jetty_websocket_jars
+  Dir.glob($PROJECT_BASE + '/jetty9-websocket/target/lib/*.jar')
 end
 
 def maven_dependency(groupid, artifactid, version, maven_repository = File.expand_path('~') + '/.m2/repository')
